move interfaces around
* Don't export interfaces that are not needed in other modules
* Move DefaultParameters to the interfaces module
* Reorder interface declarations within the modules
